Added clickdummy with user actions in the page not the toolbar
The plone.app.toolbar can add the editting toolbar, but integrating
personal tools would have two results.
Either a logged in user without editting privileges will not see he is
logged in because there is no toolbar or you will confront regular
users with an nearly empty editting toolbar which is not themed.